api-gateway/requirements.txt
# Веб-фреймворк
fastapi>=0.100.0,<0.116.0
uvicorn>=0.20.0,<0.35.0
starlette>=0.30.0,<0.47.0

# Базы данных
SQLAlchemy>=2.0.0,<2.1.0
alembic>=1.10.0,<1.15.0
psycopg2>=2.9.0,<2.10.0

# Аутентификация
passlib>=1.7.0,<1.8.0
bcrypt>=4.0.0,<4.1.0
python-jose>=3.3.0,<3.4.0
python-multipart>=0.0.6,<0.0.7

# Для работы с данными
pandas>=2.0.0,<2.3.0
numpy>=1.20.0,<2.3.0
statsmodels>=0.14.0,<0.15.0
matplotlib>=3.7.0,<3.11.0
seaborn>=0.12.0,<0.14.0

# Прочие инструменты
pydantic>=2.0.0,<2.11.0
pydantic-settings>=2.0.0,<2.1.0
python-dotenv>=1.0.0,<1.1.0
python-dateutil>=2.8.0,<2.9.0
email-validator>=2.0.0,<3.0.0
requests>=2.28.0,<2.32.0